Backend/requirements.txt
Django==3.2.10
django-cors-headers==3.13.0
django-graphql-jwt==0.3.4
graphene==3.1
graphene-django==3.0.0
mysqlclient